Refactor to avoid shared state
New approach is:

- Channels for ALL THE THINGS!!!
- Avoid almost all shared state
- Use the fancy compiler-enforced state machine thing (not really convinced it's
  worth the boilerplate, but at least it makes it a bit easier to find the right methods)

The only potential downside I could see is that it's no longer possible to
receive requests and do other work in parallel. But in JS this doesn't matter
anyways since there are no threads, and practically speaking all the mutexes
made real parallelism unlikely to begin with.